ingestion_service/requirements.txt
aiohttp
python-dotenv
pika==1.3.2
celery[redis]==5.3.6
redis==4.6.0
requests
uvicorn
fastapi
pydantic
websockets